+++ database/migrations/2024_06_07_192659_create_orders.php
@@ -27,9 +27,9 @@ return new class extends Migration
             $table->string('address')->nullable();
             $table->string('user_note')->nullable();
             $table->string('manager_note')->nullable();
-            $table->decimal('deposit_fee', 20, 8)->default(0);
-            $table->decimal('fine_fee', 20, 8)->default(0);
-            $table->decimal('total_fee', 20, 8)->default(0);
+            $table->unsignedBigInteger('deposit_fee')->default(0);
+            $table->unsignedBigInteger('fine_fee')->default(0);
+            $table->unsignedBigInteger('total_fee')->default(0);
             $table->enum('status', ['pending', 'approved', 'wating_take_book', 'hiring', 'increasing', 'wating_return', 'completed', 'canceled', 'out_of_date'])->default('pending');
             $table->timestamps();
