+++ database/migrations/2024_06_07_192706_create_order_details.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('order_details', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('order_id');
-            $table->decimal('service_fee', 20, 8)->default(0);
+            $table->unsignedBigInteger('service_fee')->default(0);
             $table->unsignedBigInteger('book_details_id');
             $table->date('return_date')->nullable();
             $table->integer('max_extensions')->default(3);
@@ -26,7 +26,7 @@ return new class extends Migration
             $table->text('comment')->nullable();
             $table->enum('status_cmt', ['hiring', 'rating_yet', 'active', 'hide', 'canceled'])->default('hiring');
             $table->enum('status_od', ['pending', 'hiring', 'completed', 'canceled', 'out_of_date'])->default('pending');
-            $table->decimal('deposit', 20, 8)->default(0);
+            $table->unsignedBigInteger('deposit')->default(0);
             $table->timestamps();
 
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
